template admin

+++ database/migrations/2025_08_03_085347_create_drainase_irigasi_pelapor_table.php
@@ -17,7 +17,7 @@ return new class extends Migration {
 			$table->text('alamat');
 			$table->string('nomor_telepon', 20);
 
-			$table->unsignedInteger('skm_id')->nullable();
+			$table->unsignedBigInteger('skm_id')->nullable();
 			$table->foreign('skm_id')->references('id')->on('skm')->onDelete('set null');
 
 			$table->timestamps();
